add comments to backend code reconfigured db files so each table is own file
