+++ app/api/chirp_routes.py
@@ -1,8 +1,7 @@
 from flask import Blueprint, jsonify, request
 from flask_login import login_required, current_user
 from app.forms.chirp_form import ChirpForm
-from app.models import db
-from app.models.db import Chirp
+from app.models import db, Chirp
 from app.s3_funcs import upload_file_to_s3, allowed_file, get_unique_filename
 
 chirp_routes = Blueprint('chirps', __name__)
@@ -17,42 +16,41 @@ def validation_errors_to_error_messages(validation_errors):
             errorMessages.append(f'{field} : {error}')
     return errorMessages
 
+
+#Get all chirps
 @chirp_routes.route('/')
 def get_chirps():
   chirps = Chirp.query.all()
   return jsonify([chirp.to_dict() for chirp in chirps])
 
+#POST a new chirp if includes image with AWS S3 functionality
 @chirp_routes.route('/new', methods=['POST'])
 def add_chirp():
 
+    #check if 'media' is in front end request take that file and set it to variable image
     if "media" not in request.files:
         return {"error": "image required"}, 400
 
     image = request.files["media"]
 
+    #confirm valid image file and send to aws s3 to get unique url to store in DB
     if not allowed_file(image.filename):
         return {"error": "file type must be png, jpg, jpeg, or gif"}, 400
 
     image.filename = get_unique_filename(image.filename)
-    print('image filename: ', image.filename)
 
     upload = upload_file_to_s3(image)
 
+    #confirm url is returned from above and return dictionary with image and user information
     if "url" not in upload:
         return upload, 400
 
     url = upload["url"]
 
-    # new_chirp = Chirp(
-    #     body=request.form.get('body'),
-    #     media=url,
-    #     userId=request.form.get('userId')
-    # )
-    # db.session.add(new_chirp)
-    # db.session.commit()
-
     return { "media": url, "user" : request.form.get("userId")}
 
+
+#POST new chirp with no image
 @chirp_routes.route('/', methods=["POST"])
 def add_chirp_noimg():
     form = ChirpForm()
@@ -68,6 +66,7 @@ def add_chirp_noimg():
     return chirp.to_dict()
 
 
+#update and existing chirp
 @chirp_routes.route('/<int:chirpId>', methods=['PUT'])
 def update_chirp(chirpId):
     chirp = Chirp.query.get(chirpId)
@@ -79,6 +78,7 @@ def update_chirp(chirpId):
 
 
 
+#delete an existing chirp
 @chirp_routes.route('/<int:chirpId>', methods=['DELETE'])
 def delete_chirp(chirpId):
     chirp = Chirp.query.get(chirpId)
@@ -88,7 +88,7 @@ def delete_chirp(chirpId):
 
 
 #like routes
-
+#like a chirp
 @chirp_routes.route('/likes/<int:chirpId>', methods=['PUT'])
 @login_required
 def likeChirp(chirpId):
@@ -97,6 +97,8 @@ def likeChirp(chirpId):
     db.session.commit()
     return chirp.to_dict()
 
+
+#unlike a chirp
 @chirp_routes.route('/unlikes/<int:chirpId>', methods=['PUT'])
 @login_required
 def unlikeChirp(chirpId):
