+++ app/models/__init__.py
@@ -1,1 +1,6 @@
-from .db import db, User, Chirp, Comment
+from .db import db
+from .chirp import Chirp
+from .comment import Comment
+from .user import User
+from .likes import likes
+from .follows import follows
